Rename some project action creators, some fixes and remove PROJECTS_CREATE action
