+++ src/__test__/components/ContentWrapper.test.jsx
@@ -15,8 +15,8 @@ import AppRouteProvider from 'utils/AppRouteProvider';
 import { makeStore } from 'redux/store';
 import { getBackendStatus } from 'redux/selectors';
 
-import { loadExperiments } from 'redux/actions/experiments';
-import { setActiveProject } from 'redux/actions/projects';
+import { loadExperiments, setActiveExperiment } from 'redux/actions/experiments';
+
 import { updateExperimentInfo } from 'redux/actions/experimentSettings';
 import generateGem2sParamsHash from 'utils/data-management/generateGem2sParamsHash';
 
@@ -120,7 +120,7 @@ describe('ContentWrapper', () => {
     }));
 
     await store.dispatch(loadExperiments());
-    await store.dispatch(setActiveProject(experimentId));
+    await store.dispatch(setActiveExperiment(experimentId));
     await store.dispatch(updateExperimentInfo({ experimentId, experimentName, sampleIds }));
   });
