+++ src/components/data-management/NewProjectModal.jsx
@@ -1,3 +1,7 @@
+import React, { useEffect, useState } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
+import { ClipLoader } from 'react-spinners';
+import PropTypes from 'prop-types';
 import {
   Button,
   Form,
@@ -6,15 +10,11 @@ import {
   Space,
   Typography,
 } from 'antd';
-import React, { useEffect, useState } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
 
-import { ClipLoader } from 'react-spinners';
-import PropTypes from 'prop-types';
-import validateInputs, { rules } from 'utils/validateInputs';
-import { createProject } from 'redux/actions/projects';
-import integrationTestConstants from 'utils/integrationTestConstants';
 import { createExperiment } from 'redux/actions/experiments';
+
+import validateInputs, { rules } from 'utils/validateInputs';
+import integrationTestConstants from 'utils/integrationTestConstants';
 
 const { Text, Title, Paragraph } = Typography;
 const { TextArea } = Input;
