+++ src/redux/reducers/projects/index.js
@@ -1,38 +1,37 @@
-import {
-  PROJECTS_CREATE,
-  PROJECTS_UPDATE,
-  PROJECTS_SET_ACTIVE,
-  PROJECTS_DELETE,
-  PROJECTS_SAVING,
-  PROJECTS_SAVED,
-  PROJECTS_ERROR,
-  PROJECTS_METADATA_CREATE,
-  PROJECTS_METADATA_UPDATE,
-  PROJECTS_METADATA_DELETE,
-  PROJECTS_LOADED,
-  PROJECTS_LOADING,
-} from '../../actionTypes/projects';
+// import {
+//   PROJECTS_CREATE,
+//   PROJECTS_UPDATE,
+//   PROJECTS_SET_ACTIVE,
+//   PROJECTS_SAVING,
+//   PROJECTS_SAVED,
+//   PROJECTS_ERROR,
+//   PROJECTS_METADATA_CREATE,
+//   PROJECTS_METADATA_UPDATE,
+//   PROJECTS_METADATA_DELETE,
+//   PROJECTS_LOADED,
+//   PROJECTS_LOADING,
+// } from '../../actionTypes/projects';
 
 import initialState from './initialState';
-import projectsCreate from './projectsCreate';
-import projectsUpdate from './projectsUpdate';
-import projectsDelete from './projectsDelete';
-import projectsSetActive from './projectsSetActive';
-import projectsSaving from './projectsSaving';
-import projectsSaved from './projectsSaved';
-import projectsError from './projectsError';
-import projectsMetadataCreate from './projectsMetadataCreate';
-import projectsMetadataUpdate from './projectsMetadataUpdate';
-import projectsMetadataDelete from './projectsMetadataDelete';
-import projectsLoaded from './projectsLoaded';
-import projectsLoading from './projectsLoading';
+// import projectsCreate from './projectsCreate';
+// import projectsUpdate from './projectsUpdate';
+// import projectsDelete from './projectsDelete';
+// import projectsSetActive from './projectsSetActive';
+// import projectsSaving from './projectsSaving';
+// import projectsSaved from './projectsSaved';
+// import projectsError from './projectsError';
+// import projectsMetadataCreate from './projectsMetadataCreate';
+// import projectsMetadataUpdate from './projectsMetadataUpdate';
+// import projectsMetadataDelete from './projectsMetadataDelete';
+// import projectsLoaded from './projectsLoaded';
+// import projectsLoading from './projectsLoading';
 
-import {
-  SAMPLES_CREATE, SAMPLES_DELETE,
-} from '../../actionTypes/samples';
+// import {
+//   SAMPLES_CREATE, SAMPLES_DELETE,
+// } from '../../actionTypes/samples';
 
-import samplesCreate from './samplesCreate';
-import samplesDelete from './samplesDelete';
+// import samplesCreate from './samplesCreate';
+// import samplesDelete from './samplesDelete';
 
 const projectsReducer = (state = initialState, action) => {
   switch (action.type) {
@@ -46,10 +45,6 @@ const projectsReducer = (state = initialState, action) => {
 
     // case PROJECTS_SET_ACTIVE: {
     //   return projectsSetActive(state, action);
-    // }
-
-    // case PROJECTS_DELETE: {
-    //   return projectsDelete(state, action);
     // }
 
     // case PROJECTS_SAVING: {
